+++ src/app/features/products/components/product-list/product-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { CartService } from 'src/app/features/cart/services/cart.service';
 import { ProductModel } from '../../models/product';
 import { ProductsService } from '../../services/products.service';
 
@@ -11,10 +12,20 @@ export class ProductListComponent implements OnInit {
 
   products: ProductModel[];
 
-  constructor(private productsService: ProductsService){}
+  constructor(
+    private productsService: ProductsService,
+    private cartService: CartService,
+  ) { }
 
 
   ngOnInit(): void {
     this.products = this.productsService.getProducts();
+  }
+
+  onAddToCart(productId: number): void {
+    const product = this.products.find(x => x.id == productId);
+    if (product) {
+      this.cartService.buyProduct(product);
+    }
   }
 }